fix(battery): dimensiones de caja a decimal (rompian el POST con 400)
Box width/length/height pasan de int? a decimal? (numeric(6,2)): las medidas pueden tener decimales (ej 17.5 cm). Antes enviar un decimal a un int? abortaba la deserializacion del comando -> 400. Migracion AlterColumn (sin perdida, columnas en null).

+++ MyCarBE.Application/Features/InspectionReports/Commands/ProposedItemInputs.cs
@@ -61,8 +61,8 @@ public record BatteryInspectionInput(
     DateOnly?     InstalledOn,
     string?       Notes,
     int?                 CapacityAh           = null,
-    int?                 BoxWidthCm           = null,
-    int?                 BoxLengthCm          = null,
-    int?                 BoxHeightCm          = null,
+    decimal?             BoxWidthCm           = null,
+    decimal?             BoxLengthCm          = null,
+    decimal?             BoxHeightCm          = null,
     BatteryTerminalSide? PositiveTerminalSide = null
 );
